+++ Sources/SuggestionEngine/models/Goal.swift
@@ -1,0 +1,14 @@
+//
+//  Goal.swift
+//  ScheduleEngine
+//
+//  Created by Rapha Vidal on 06/10/25.
+//
+import Foundation
+
+// MARK: - Enums
+
+// Objetivos
+public enum Goal: String, Codable, Sendable {
+    case loseWeight, gainMuscle, improveFitness, IncreaseFlexibilityAndMobility, healthAndWellBeing, socialization
+}
