+++ Sources/SuggestionEngine/models/Location.swift
@@ -1,0 +1,14 @@
+//
+//  Location.swift
+//  ScheduleEngine
+//
+//  Created by Rapha Vidal on 06/10/25.
+//
+import Foundation
+
+// MARK: - Enums
+
+// Locais
+public enum Location: String, Codable, Sendable {
+    case home, gym, work, outdoor
+}
